fixing duplicated form key

+++ lib/features/auth/controller/auth_controller.dart
@@ -145,7 +145,8 @@ class AuthController extends GetxController {
 
   void clearAllAuthData() {
     LocalStorageKey.values.forEach((key) {
-      LocalStorage.deleteDataUnderKey(key);
+      if (key != LocalStorageKey.firstOpen)
+        LocalStorage.deleteDataUnderKey(key);
     });
   }
